tools/nut-scanner/nut-scan.h + scan_xml_http.c: fix port type to uint16_t

+++ tools/nut-scanner/nut-scan.h
@@ -113,8 +113,8 @@ typedef struct nutscan_ipmi {
 
 /* XML HTTP structure */
 typedef struct nutscan_xml {
-	int port_http;		/* Port for xml http (tcp) */
-	int port_udp;		/* Port for xml udp */
+	uint16_t port_http;		/* Port for xml http (tcp) */
+	uint16_t port_udp;		/* Port for xml udp */
 	long usec_timeout;	/* Wait this long for a response */
 	char *peername;		/* Hostname or NULL for broadcast mode */
 } nutscan_xml_t;
